+++ NEWS.md
@@ -1,6 +1,7 @@
 # mvgam 1.1.5 (not yet on CRAN)
 ## New functionalities
 * Converted several more plotting functions to return `ggplot` objects in place of base R plots
+* Added four new `type`s to the `pp_check()` function to allow more targeted investigations of randomized quantile residual distributions
 
 ## Bug fixes
 * Bug fix to ensure piecewise trends are extrapolated the correct number of timepoints when forecasting using the `forecast()` function